pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scrape-stream"
version = "0.1.0"
description = "DI.FM stream ripper"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tqdm>=4.67.2",
    "mutagen>=1.47.0",
]

[project.scripts]
scrape = "scrape_stream:main"
serve = "serve_stream:main"

[tool.uv]
package = true
